Conflicts:
	org.hive2hive.client/pom.xml
	org.hive2hive.client/src/main/resources/logback.xml
	org.hive2hive.core/pom.xml
	org.hive2hive.core/src/main/java/org/hive2hive/core/exceptions/AbortModificationCode.java

+++ org.hive2hive.core/src/test/java/org/hive2hive/core/processes/share/ShareFolderTest.java
@@ -2,11 +2,10 @@ package org.hive2hive.core.processes.share;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
+import java.util.List;
 import java.util.Random;
 
 import org.apache.commons.io.FileUtils;
-import org.hive2hive.core.H2HConstants;
 import org.hive2hive.core.H2HJUnitTest;
 import org.hive2hive.core.events.framework.interfaces.file.IFileAddEvent;
 import org.hive2hive.core.events.framework.interfaces.file.IFileShareEvent;
@@ -34,7 +33,7 @@ import org.junit.Test;
 public class ShareFolderTest extends H2HJUnitTest {
 
 	private static final int networkSize = 6;
-	private static ArrayList<NetworkManager> network;
+	private static List<NetworkManager> network;
 
 	private static File rootA;
 	private static File rootB;
@@ -66,24 +65,21 @@ public class ShareFolderTest extends H2HJUnitTest {
 	}
 
 	@Test
-	public void shareFilledFolderTest() throws IOException, IllegalArgumentException, NoSessionException, GetFailedException,
-			InterruptedException, NoPeerConnectionException {
+	public void shareFilledFolderTest() throws IOException, IllegalArgumentException, NoSessionException,
+			GetFailedException, InterruptedException, NoPeerConnectionException {
 		// upload an empty folder
 		File folderToShare = new File(rootA, "sharedFolder");
 		folderToShare.mkdirs();
 		UseCaseTestUtil.uploadNewFile(network.get(0), folderToShare);
 
-		File file1 = FileTestUtil.createFileRandomContent("file1", new Random().nextInt(5) + 1, folderToShare,
-				H2HConstants.DEFAULT_CHUNK_SIZE);
+		File file1 = FileTestUtil.createFileRandomContent("file1", new Random().nextInt(5) + 1, folderToShare);
 		UseCaseTestUtil.uploadNewFile(network.get(0), file1);
-		File file2 = FileTestUtil.createFileRandomContent("file2", new Random().nextInt(5) + 1, folderToShare,
-				H2HConstants.DEFAULT_CHUNK_SIZE);
+		File file2 = FileTestUtil.createFileRandomContent("file2", new Random().nextInt(5) + 1, folderToShare);
 		UseCaseTestUtil.uploadNewFile(network.get(0), file2);
 		File subfolder = new File(folderToShare, "subfolder1");
 		subfolder.mkdir();
 		UseCaseTestUtil.uploadNewFile(network.get(0), subfolder);
-		File file3 = FileTestUtil.createFileRandomContent("file3", new Random().nextInt(5) + 1, subfolder,
-				H2HConstants.DEFAULT_CHUNK_SIZE);
+		File file3 = FileTestUtil.createFileRandomContent("file3", new Random().nextInt(5) + 1, subfolder);
 		UseCaseTestUtil.uploadNewFile(network.get(0), file3);
 
 		// share the filled folder
@@ -154,8 +150,12 @@ public class ShareFolderTest extends H2HJUnitTest {
 
 	@AfterClass
 	public static void endTest() throws IOException {
-		FileUtils.deleteDirectory(rootA);
-		FileUtils.deleteDirectory(rootB);
+		if (rootA.exists()) {
+			FileUtils.deleteDirectory(rootA);
+		}
+		if (rootB.exists()) {
+			FileUtils.deleteDirectory(rootB);
+		}
 		NetworkTestUtil.shutdownNetwork(network);
 		afterClass();
 	}
